moved context_sampler.py
